+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### 🐞 Fixed
 - `Channel.update` unnecessarily causing an actual name and imageURL update on 1-1 channels [#973](https://github.com/GetStream/stream-chat-swift/issues/973)
+- `Channel.membership` is `nil` for when channel is obtained from `presenter.rx.channelDidUpdate` observable
 
 # [2.6.6](https://github.com/GetStream/stream-chat-swift/releases/tag/2.6.6)
 _May 04, 2021_