+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Upcoming
 
+### 🔄 Changed
+
+# [2.6.7](https://github.com/GetStream/stream-chat-swift/releases/tag/2.6.7)
+_May 18, 2021_
+
 ### ⚠️ Deprecated
 - `Client.shared.update(channel:name:imageURL:extraData:_:)` and `channel.update(name:imageURL:extraData:_:)` is deprecated. Please use versions without `name:imageURL:`